+++ rustfst-cli/src/cmds/connect.rs
@@ -1,13 +1,16 @@
 use rustfst::prelude::*;
 
-use log::info;
+use log::debug;
 
 use failure::Fallible;
 
 pub fn connect_cli(path_in: &str, path_out: &str) -> Fallible<()> {
-    info!("Connect");
+    debug!("Reading FST");
     let mut fst = VectorFst::<TropicalWeight>::read(path_in)?;
+    debug!("Running Connect algorithm");
     connect(&mut fst)?;
+    debug!("Writing FST");
     fst.write(path_out)?;
+    debug!("Done");
     Ok(())
 }