+++ rustfst-cli/src/cmds/mod.rs
@@ -1,0 +1,6 @@
+pub mod arcsort;
+pub mod connect;
+pub mod invert;
+pub mod minimize;
+pub mod project;
+pub mod topsort;